requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
